feat: ProductService 인터페이스/구현에 검색 기반 메서드 추가
